+++ packages/docs/src/stories/Box.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { Box, BoxProps } from '@marcosvinicius-design-system/react'
+import { Box, BoxProps } from '@marcosvinicius_design-system/react'
 
 export default {
   title: 'Surface/Box',