+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @marcosvinicius_design-system/react
 
+## 2.0.0
+
+### Major Changes
+
+- 3be2bf6: summary
+
 ## 1.0.2
 
 ### Patch Changes